requirements.txt
arabic_reshaper==2.1.3
imutils==0.5.4
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.1.1
pydantic==1.9.0
pytesseract==0.3.9
python_bidi==0.4.2
scikit_image==0.19.2
skimage==0.0
